Rename components directory
