+++ src/components/WithAuthentication.tsx
